Move some io functions into io module
