+++ src/io/free_functions.rs
@@ -1,16 +1,18 @@
 use std::fs::File;
-use std::io::{BufRead, BufWriter, Seek};
-use std::iter;
+use std::io::{self, BufRead, BufWriter, Seek, Write as _};
 use std::path::Path;
+use std::{iter, mem};
 
 use crate::{codecs::*, ExtendedColorType, ImageReader};
 
-use crate::DynamicImage;
-use crate::error::{ImageError, ImageFormatHint, ImageResult};
-use crate::error::{UnsupportedError, UnsupportedErrorKind};
+use crate::error::{
+    ImageError, ImageFormatHint, ImageResult, LimitError, LimitErrorKind, ParameterError,
+    ParameterErrorKind, UnsupportedError, UnsupportedErrorKind,
+};
 use crate::image::ImageFormat;
 #[allow(unused_imports)] // When no features are supported
 use crate::image::{ImageDecoder, ImageEncoder};
+use crate::DynamicImage;
 
 /// Create a new image from a Reader.
 ///
@@ -65,7 +67,7 @@ pub fn save_buffer_with_format(
 
 #[allow(unused_variables)]
 // Most variables when no features are supported
-pub(crate) fn write_buffer_impl<W: std::io::Write + Seek>(
+pub(crate) fn write_buffer_impl<W: io::Write + Seek>(
     buffered_write: &mut W,
     buf: &[u8],
     width: u32,
@@ -197,4 +199,345 @@ pub(crate) fn guess_format_impl(buffer: &[u8]) -> Option<ImageFormat> {
     }
 
     None
+}
+
+/// Decodes a specific region of the image, represented by the rectangle
+/// starting from ```x``` and ```y``` and having ```length``` and ```width```
+#[allow(dead_code)]
+// When no image formats that use it are enabled
+pub(crate) fn load_rect<D, F1, F2, E>(
+    x: u32,
+    y: u32,
+    width: u32,
+    height: u32,
+    buf: &mut [u8],
+    row_pitch: usize,
+    decoder: &mut D,
+    scanline_bytes: usize,
+    mut seek_scanline: F1,
+    mut read_scanline: F2,
+) -> ImageResult<()>
+where
+    D: ImageDecoder,
+    F1: FnMut(&mut D, u64) -> io::Result<()>,
+    F2: FnMut(&mut D, &mut [u8]) -> Result<(), E>,
+    ImageError: From<E>,
+{
+    let scanline_bytes = u64::try_from(scanline_bytes).unwrap();
+    let row_pitch = u64::try_from(row_pitch).unwrap();
+
+    let (x, y, width, height) = (
+        u64::from(x),
+        u64::from(y),
+        u64::from(width),
+        u64::from(height),
+    );
+    let dimensions = decoder.dimensions();
+    let bytes_per_pixel = u64::from(decoder.color_type().bytes_per_pixel());
+    let row_bytes = bytes_per_pixel * u64::from(dimensions.0);
+    let total_bytes = width * height * bytes_per_pixel;
+
+    assert!(
+        buf.len() >= usize::try_from(total_bytes).unwrap_or(usize::MAX),
+        "output buffer too short\n expected `{}`, provided `{}`",
+        total_bytes,
+        buf.len()
+    );
+
+    let mut current_scanline = 0;
+    let mut tmp = Vec::new();
+    let mut tmp_scanline = None;
+
+    {
+        // Read a range of the image starting from byte number `start` and continuing until byte
+        // number `end`. Updates `current_scanline` and `bytes_read` appropriately.
+        let mut read_image_range =
+            |mut start: u64, end: u64, mut output: &mut [u8]| -> ImageResult<()> {
+                // If the first scanline we need is already stored in the temporary buffer, then handle
+                // it first.
+                let target_scanline = start / scanline_bytes;
+                if tmp_scanline == Some(target_scanline) {
+                    let position = target_scanline * scanline_bytes;
+                    let offset = start.saturating_sub(position);
+                    let len = (end - start)
+                        .min(scanline_bytes - offset)
+                        .min(end - position);
+
+                    output
+                        .write_all(&tmp[offset as usize..][..len as usize])
+                        .unwrap();
+                    start += len;
+
+                    if start == end {
+                        return Ok(());
+                    }
+                }
+
+                let target_scanline = start / scanline_bytes;
+                if target_scanline != current_scanline {
+                    seek_scanline(decoder, target_scanline)?;
+                    current_scanline = target_scanline;
+                }
+
+                let mut position = current_scanline * scanline_bytes;
+                while position < end {
+                    if position >= start && end - position >= scanline_bytes {
+                        read_scanline(decoder, &mut output[..(scanline_bytes as usize)])?;
+                        output = &mut output[scanline_bytes as usize..];
+                    } else {
+                        tmp.resize(scanline_bytes as usize, 0u8);
+                        read_scanline(decoder, &mut tmp)?;
+                        tmp_scanline = Some(current_scanline);
+
+                        let offset = start.saturating_sub(position);
+                        let len = (end - start)
+                            .min(scanline_bytes - offset)
+                            .min(end - position);
+
+                        output
+                            .write_all(&tmp[offset as usize..][..len as usize])
+                            .unwrap();
+                    }
+
+                    current_scanline += 1;
+                    position += scanline_bytes;
+                }
+                Ok(())
+            };
+
+        if x + width > u64::from(dimensions.0)
+            || y + height > u64::from(dimensions.1)
+            || width == 0
+            || height == 0
+        {
+            return Err(ImageError::Parameter(ParameterError::from_kind(
+                ParameterErrorKind::DimensionMismatch,
+            )));
+        }
+        if scanline_bytes > usize::MAX as u64 {
+            return Err(ImageError::Limits(LimitError::from_kind(
+                LimitErrorKind::InsufficientMemory,
+            )));
+        }
+
+        if x == 0 && width == u64::from(dimensions.0) && row_pitch == row_bytes {
+            let start = x * bytes_per_pixel + y * row_bytes;
+            let end = (x + width) * bytes_per_pixel + (y + height - 1) * row_bytes;
+            read_image_range(start, end, buf)?;
+        } else {
+            for (output_slice, row) in buf.chunks_mut(row_pitch as usize).zip(y..(y + height)) {
+                let start = x * bytes_per_pixel + row * row_bytes;
+                let end = (x + width) * bytes_per_pixel + row * row_bytes;
+                read_image_range(start, end, output_slice)?;
+            }
+        }
+    }
+
+    // Seek back to the start
+    Ok(seek_scanline(decoder, 0)?)
+}
+
+/// Reads all of the bytes of a decoder into a Vec<T>. No particular alignment
+/// of the output buffer is guaranteed.
+///
+/// Panics if there isn't enough memory to decode the image.
+pub(crate) fn decoder_to_vec<T>(decoder: impl ImageDecoder) -> ImageResult<Vec<T>>
+where
+    T: crate::traits::Primitive + bytemuck::Pod,
+{
+    let total_bytes = usize::try_from(decoder.total_bytes());
+    if total_bytes.is_err() || total_bytes.unwrap() > isize::MAX as usize {
+        return Err(ImageError::Limits(LimitError::from_kind(
+            LimitErrorKind::InsufficientMemory,
+        )));
+    }
+
+    let mut buf = vec![num_traits::Zero::zero(); total_bytes.unwrap() / mem::size_of::<T>()];
+    decoder.read_image(bytemuck::cast_slice_mut(buf.as_mut_slice()))?;
+    Ok(buf)
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::ColorType;
+    use std::io;
+
+    use super::{load_rect, ImageDecoder, ImageResult};
+
+    #[test]
+    fn test_load_rect() {
+        struct MockDecoder {
+            scanline_number: u64,
+            scanline_bytes: u64,
+        }
+        impl ImageDecoder for MockDecoder {
+            fn dimensions(&self) -> (u32, u32) {
+                (5, 5)
+            }
+            fn color_type(&self) -> ColorType {
+                ColorType::L8
+            }
+            fn read_image(self, _buf: &mut [u8]) -> ImageResult<()> {
+                unimplemented!()
+            }
+            fn read_image_boxed(self: Box<Self>, buf: &mut [u8]) -> ImageResult<()> {
+                (*self).read_image(buf)
+            }
+        }
+
+        const DATA: [u8; 25] = [
+            0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23,
+            24,
+        ];
+
+        fn seek_scanline(m: &mut MockDecoder, n: u64) -> io::Result<()> {
+            m.scanline_number = n;
+            Ok(())
+        }
+        fn read_scanline(m: &mut MockDecoder, buf: &mut [u8]) -> io::Result<()> {
+            let bytes_read = m.scanline_number * m.scanline_bytes;
+            if bytes_read >= 25 {
+                return Ok(());
+            }
+
+            let len = m.scanline_bytes.min(25 - bytes_read);
+            buf[..(len as usize)].copy_from_slice(&DATA[(bytes_read as usize)..][..(len as usize)]);
+            m.scanline_number += 1;
+            Ok(())
+        }
+
+        for scanline_bytes in 1..30 {
+            let mut output = [0u8; 26];
+
+            load_rect(
+                0,
+                0,
+                5,
+                5,
+                &mut output,
+                5,
+                &mut MockDecoder {
+                    scanline_number: 0,
+                    scanline_bytes,
+                },
+                scanline_bytes as usize,
+                seek_scanline,
+                read_scanline,
+            )
+            .unwrap();
+            assert_eq!(output[0..25], DATA);
+            assert_eq!(output[25], 0);
+
+            output = [0u8; 26];
+            load_rect(
+                3,
+                2,
+                1,
+                1,
+                &mut output,
+                1,
+                &mut MockDecoder {
+                    scanline_number: 0,
+                    scanline_bytes,
+                },
+                scanline_bytes as usize,
+                seek_scanline,
+                read_scanline,
+            )
+            .unwrap();
+            assert_eq!(output[0..2], [13, 0]);
+
+            output = [0u8; 26];
+            load_rect(
+                3,
+                2,
+                2,
+                2,
+                &mut output,
+                2,
+                &mut MockDecoder {
+                    scanline_number: 0,
+                    scanline_bytes,
+                },
+                scanline_bytes as usize,
+                seek_scanline,
+                read_scanline,
+            )
+            .unwrap();
+            assert_eq!(output[0..5], [13, 14, 18, 19, 0]);
+
+            output = [0u8; 26];
+            load_rect(
+                1,
+                1,
+                2,
+                4,
+                &mut output,
+                2,
+                &mut MockDecoder {
+                    scanline_number: 0,
+                    scanline_bytes,
+                },
+                scanline_bytes as usize,
+                seek_scanline,
+                read_scanline,
+            )
+            .unwrap();
+            assert_eq!(output[0..9], [6, 7, 11, 12, 16, 17, 21, 22, 0]);
+        }
+    }
+
+    #[test]
+    fn test_load_rect_single_scanline() {
+        const DATA: [u8; 25] = [
+            0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23,
+            24,
+        ];
+
+        struct MockDecoder;
+        impl ImageDecoder for MockDecoder {
+            fn dimensions(&self) -> (u32, u32) {
+                (5, 5)
+            }
+            fn color_type(&self) -> ColorType {
+                ColorType::L8
+            }
+            fn read_image(self, _buf: &mut [u8]) -> ImageResult<()> {
+                unimplemented!()
+            }
+            fn read_image_boxed(self: Box<Self>, buf: &mut [u8]) -> ImageResult<()> {
+                (*self).read_image(buf)
+            }
+        }
+
+        // Ensure that seek scanline is called only once.
+        let mut seeks = 0;
+        let seek_scanline = |_d: &mut MockDecoder, n: u64| -> io::Result<()> {
+            seeks += 1;
+            assert_eq!(n, 0);
+            assert_eq!(seeks, 1);
+            Ok(())
+        };
+
+        fn read_scanline(_m: &mut MockDecoder, buf: &mut [u8]) -> io::Result<()> {
+            buf.copy_from_slice(&DATA);
+            Ok(())
+        }
+
+        let mut output = [0; 26];
+        load_rect(
+            1,
+            1,
+            2,
+            4,
+            &mut output,
+            2,
+            &mut MockDecoder,
+            DATA.len(),
+            seek_scanline,
+            read_scanline,
+        )
+        .unwrap();
+        assert_eq!(output[0..9], [6, 7, 11, 12, 16, 17, 21, 22, 0]);
+    }
 }
